Chapter3/Framework_pipeline/pipeline/3_train/requirements.txt
click
dill
tensorflow==2.3.0
google-cloud-storage
pandas
pillow
sklearn